refactor(infra): rewrite imports to infra namespaces
