--- src-tauri/src/adapters.rs
@@ -1,1 +1,0 @@
-pub use crate::infra::adapters::*;
