+++ src-tauri/src/infra/mod.rs
@@ -1,8 +1,8 @@
 pub mod adapters;
-mod adapter_registry;
 mod mutation;
 pub mod parsers;
+pub mod registry;
 pub mod security;
 
-pub use adapter_registry::AdapterRegistry;
+pub use registry::AdapterRegistry;
 pub use mutation::{MutationTestHooks, SafeFileMutator};
