+++ src-tauri/src/infra/registry/mod.rs
@@ -1,0 +1,3 @@
+mod adapter_registry;
+
+pub use adapter_registry::AdapterRegistry;
