--- src-tauri/src/parsers.rs
@@ -1,1 +1,0 @@
-pub use crate::infra::parsers::*;
