--- src-tauri/src/security.rs
@@ -1,1 +1,0 @@
-pub use crate::infra::security::*;
